complete Fragment Object States and Transaction
Extremely helpful app in understanding the different states of fragment
objects.. eg: attached, detached, etc

Note thansaction.detach(frag) dows not really detach the frag object
from its activity.. rather, it just pauses and stops the frag, thus
making it invisible for the user.. the object "appears" to be detached,
but the onDetach() method is NOT called when transaction.detach(frag) is
used. Thus, the activity is fully aware of the existence of the fragment
object, and the link between the two is not lost.